+++ changelog.md
@@ -62,6 +62,7 @@ Codion Change Log
 - EntityTablePanel.Config.configureTable() added, createTableCellRenderer() removed.
 - TabbedDetailLayout, expand/collapse improved, now stops in the middle when expanding/collapsing from the opposite extreme.
 - TabbedDetailLayout now activates parent panel when detail panel is collapsed.
+- EntityEditPanel, crud operation builders added, EntityDialogs refactored.
 
 ## 0.17.43
 ### is.codion.common.core